Update header or something

+++ song.h
@@ -1,7 +1,7 @@
 #include <stdint.h>
 
 struct Note {
-	int delta_time;
+	uint8_t delta_time;
 	uint8_t note;	
 };
